move Search module into Rsbe::Client namespace
