+++ lib/rsbe/client.rb
@@ -7,7 +7,8 @@ se
 connection not_found_error
 method_not_implemented_error
 wrong_origin_error
-unrecognized_resource_error).each do |r|
+unrecognized_resource_error
+search).each do |r|
   require_relative "./client/#{r}"
 end
